+++ pqc/src/certificate/pq_certificate.h
@@ -1,0 +1,6 @@
+#ifndef PQ_CERTIFICATE_H
+#define PQ_CERTIFICATE_H
+
+#include "pq_certificate_funcs.h"
+
+#endif
