+++ pqc/src/certificate/pq_certificate_params.h
@@ -1,0 +1,11 @@
+#ifndef PQ_CERTIFICATE_PARAMS_H
+#define PQ_CERTIFICATE_PARAMS_H
+
+#include <openssl/evp.h>
+#include <openssl/pem.h>
+#include <openssl/x509.h>
+#include <openssl/x509v3.h>
+#include <openssl/ec.h>
+#include <openssl/err.h>
+
+#endif
