+++ pqc/src/client/pq_client_params.h
@@ -1,8 +1,12 @@
+#ifndef PQ_CLIENT_PARAMS_H
+#define PQ_CLIENT_PARAMS_H
 
-#include "pq_openssl.h"
+#include "../pq_openssl.h"
 
 // Server's IP address
 #define SERVER_ADDR "0.0.0.0"
 #define SERVER_PORT 1316
 
 #define BUFFER_SIZE 1024
+
+#endif
