--- pqc/src/pq_server.h
@@ -1,5 +1,0 @@
-
-#include "pq_server_funcs.h"
-
-void run_server();
-int main();
