+++ pqc/src/server/pq_server.h
@@ -1,0 +1,9 @@
+#ifndef PQ_SERVER_H
+#define PQ_SERVER_H
+
+#include "pq_server_funcs.h"
+
+void run_server();
+int main();
+
+#endif
